Now JWTtoken being checked well in Users endpoints

+++ src/main/java/hu/bme/aut/shed/controller/UserController.java
@@ -1,5 +1,6 @@
 package hu.bme.aut.shed.controller;
 
+import hu.bme.aut.shed.component.JwtTokenUtil;
 import hu.bme.aut.shed.dto.Request.OtpRequest;
 import hu.bme.aut.shed.exception.UserNotFoundException;
 import hu.bme.aut.shed.model.User;
@@ -9,6 +10,7 @@ import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.AuthorizationServiceException;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
@@ -21,24 +23,32 @@ public class UserController {
     @Autowired
     private OtpService otpService;
 
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
+
     @RequestMapping(value = "/", method = {RequestMethod.GET}, produces = "application/json")
-    public ResponseEntity<User> getUserById(@RequestParam Long id) {
+    public ResponseEntity<?> getUserById(@RequestParam Long id) {
         try {
             User user = userService.getById(id);
             return new ResponseEntity<>(user, HttpStatus.OK);
         }
         catch (UserNotFoundException exception){
-            return new ResponseEntity<>(null, HttpStatus.NOT_FOUND);
+            return new ResponseEntity<>(exception.getMessage(), HttpStatus.NOT_FOUND);
         }
     }
 
     @RequestMapping(value = "/edit/", method = {RequestMethod.PUT}, produces = "application/json")
-    public ResponseEntity<User> updateUser(@RequestParam Long id,@RequestBody User editedUser) {
+    public ResponseEntity<?> updateUser(@RequestParam Long id,@RequestBody User editedUser) {
         try{
+            String token = JwtTokenUtil.getToken();
+            User currentUser = jwtTokenUtil.getUserFromToken(token);
+            if (!editedUser.getId().equals(currentUser.getId())){
+                    throw new AuthorizationServiceException("You dont have permission to make changes");
+            }
             User user = userService.updateById(id,editedUser);
             return new ResponseEntity<>(user, HttpStatus.OK);
         } catch (Exception e) {
-            return new ResponseEntity<>(null,HttpStatus.BAD_REQUEST);
+            return new ResponseEntity<>(e.getMessage(),HttpStatus.BAD_REQUEST);
         }
     }
 
@@ -57,12 +67,17 @@ public class UserController {
     }
 
     @RequestMapping(value = "/delete/", method = {RequestMethod.DELETE}, produces = "application/json")
-    public ResponseEntity<HttpStatus> deleteUser(@RequestParam Long id) {
+    public ResponseEntity<?> deleteUser(@RequestParam Long id) {
         try {
+            String token = JwtTokenUtil.getToken();
+            User currentUser = jwtTokenUtil.getUserFromToken(token);
+            if (userService.getById(id).getId().equals(currentUser.getId())){
+                throw new AuthorizationServiceException("You dont have permission to make changes");
+            }
             userService.deleteById(id);
             return new ResponseEntity<>(HttpStatus.NO_CONTENT);
         } catch (Exception e) {
-            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+            return new ResponseEntity<>(e.getMessage(),HttpStatus.INTERNAL_SERVER_ERROR);
         }
     }
